Refactor : JWT 발급 로직 리팩토링
