+++ src/main/java/com/example/malang/config/SecurityFilterBeanConfig.java
@@ -1,15 +1,11 @@
 package com.example.malang.config;
 
-import com.example.malang.oauth.filter.JwtAuthorizationFilter;
-//import com.example.malang.oauth.handler.JwtProviderHandler;
-import com.example.malang.oauth.service.JwtService;
+import com.example.malang.jwt.JwtAuthorizationFilter;
+import com.example.malang.jwt.JwtService;
 import com.example.malang.repository.MemberRepository;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
 
 @Configuration
 @RequiredArgsConstructor
