+++ src/main/java/com/example/malang/dto/MemberRequestDto.java
@@ -14,6 +14,4 @@ public class MemberRequestDto {
         private String email;
         private String memberName;
     }
-
-
 }
