+++ src/main/java/com/example/malang/jwt/JwtService.java
@@ -1,16 +1,12 @@
-package com.example.malang.oauth.service;
+package com.example.malang.jwt;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
-import com.auth0.jwt.interfaces.DecodedJWT;
 import com.example.malang.exception.BaseException;
 import com.example.malang.exception.ErrorCode;
-import com.example.malang.oauth.common.TokenMapping;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
