+++ src/main/java/com/example/malang/jwt/TokenMapping.java
@@ -1,8 +1,7 @@
-package com.example.malang.oauth.common;
+package com.example.malang.jwt;
 
 import lombok.Builder;
 import lombok.Getter;
-import org.springframework.web.bind.annotation.GetMapping;
 
 @Getter
 public class TokenMapping {
